* all pass ok * fix code_style test=develop

+++ lite/api/test_resnet50_lite_bm.cc
@@ -107,7 +107,7 @@ void TestModel(const std::vector<Place>& valid_places) {
 
 TEST(ResNet50, test_bm) {
   std::vector<Place> valid_places({
-      Place{TARGET(kBM), PRECISION(kInt8)}
+      Place{TARGET(kBM), PRECISION(kFloat)}
   });
 
   TestModel(valid_places);
